codex-tools: extract discoverable tool models
