+++ codex-rs/core/src/plugins/discoverable.rs
@@ -9,6 +9,7 @@ use super::PluginsManager;
 use crate::config::Config;
 use crate::config::types::ToolSuggestDiscoverableType;
 use codex_features::Feature;
+use codex_tools::DiscoverablePluginInfo;
 
 const TOOL_SUGGEST_DISCOVERABLE_PLUGIN_ALLOWLIST: &[&str] = &[
     "github@openai-curated",
@@ -23,7 +24,7 @@ const TOOL_SUGGEST_DISCOVERABLE_PLUGIN_ALLOWLIST: &[&str] = &[
 
 pub(crate) fn list_tool_suggest_discoverable_plugins(
     config: &Config,
-) -> anyhow::Result<Vec<PluginCapabilitySummary>> {
+) -> anyhow::Result<Vec<DiscoverablePluginInfo>> {
     if !config.features.enabled(Feature::Plugins) {
         return Ok(Vec::new());
     }
@@ -47,7 +48,7 @@ pub(crate) fn list_tool_suggest_discoverable_plugins(
         return Ok(Vec::new());
     };
 
-    let mut discoverable_plugins = Vec::<PluginCapabilitySummary>::new();
+    let mut discoverable_plugins = Vec::<DiscoverablePluginInfo>::new();
     for plugin in curated_marketplace.plugins {
         if plugin.installed
             || (!TOOL_SUGGEST_DISCOVERABLE_PLUGIN_ALLOWLIST.contains(&plugin.id.as_str())
@@ -66,14 +67,28 @@ pub(crate) fn list_tool_suggest_discoverable_plugins(
                 marketplace_path: curated_marketplace.path.clone(),
             },
         ) {
-            Ok(plugin) => discoverable_plugins.push(plugin.plugin.into()),
+            Ok(plugin) => {
+                let plugin: PluginCapabilitySummary = plugin.plugin.into();
+                discoverable_plugins.push(DiscoverablePluginInfo {
+                    id: plugin.config_name,
+                    name: plugin.display_name,
+                    description: plugin.description,
+                    has_skills: plugin.has_skills,
+                    mcp_server_names: plugin.mcp_server_names,
+                    app_connector_ids: plugin
+                        .app_connector_ids
+                        .into_iter()
+                        .map(|connector_id| connector_id.0)
+                        .collect(),
+                });
+            }
             Err(err) => warn!("failed to load discoverable plugin suggestion {plugin_id}: {err:#}"),
         }
     }
     discoverable_plugins.sort_by(|left, right| {
-        left.display_name
-            .cmp(&right.display_name)
-            .then_with(|| left.config_name.cmp(&right.config_name))
+        left.name
+            .cmp(&right.name)
+            .then_with(|| left.id.cmp(&right.id))
     });
     Ok(discoverable_plugins)
 }
