+++ codex-rs/tools/Cargo.toml
@@ -8,6 +8,7 @@ version.workspace = true
 workspace = true
 
 [dependencies]
+codex-app-server-protocol = { workspace = true }
 codex-code-mode = { workspace = true }
 codex-protocol = { workspace = true }
 rmcp = { workspace = true, default-features = false, features = [
