+++ codex-rs/tools/README.md
@@ -27,7 +27,7 @@ schema and Responses API tool primitives that no longer need to live in
 - local host tool spec builders for shell/exec/request-permissions/view-image
 - collaboration and agent-job `ToolSpec` builders for spawn/send/wait/close,
   `request_user_input`, and CSV fanout/reporting
-- tool discovery and suggestion models / `ToolSpec` builders for
+- discoverable-tool models, client filtering, and `ToolSpec` builders for
   `tool_search` and `tool_suggest`
 - `parse_tool_input_schema()`
 - `parse_dynamic_tool()`
