+++ codex-rs/tools/src/tool_discovery.rs
@@ -1,9 +1,12 @@
 use crate::JsonSchema;
 use crate::ResponsesApiTool;
 use crate::ToolSpec;
+use codex_app_server_protocol::AppInfo;
 use serde::Deserialize;
 use serde::Serialize;
 use std::collections::BTreeMap;
+
+const TUI_CLIENT_NAME: &str = "codex-tui";
 
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub struct ToolSearchAppInfo {
@@ -32,6 +35,78 @@ impl DiscoverableToolType {
 pub enum DiscoverableToolAction {
     Install,
     Enable,
+}
+
+#[derive(Clone, Debug, PartialEq)]
+pub enum DiscoverableTool {
+    Connector(Box<AppInfo>),
+    Plugin(Box<DiscoverablePluginInfo>),
+}
+
+impl DiscoverableTool {
+    pub fn tool_type(&self) -> DiscoverableToolType {
+        match self {
+            Self::Connector(_) => DiscoverableToolType::Connector,
+            Self::Plugin(_) => DiscoverableToolType::Plugin,
+        }
+    }
+
+    pub fn id(&self) -> &str {
+        match self {
+            Self::Connector(connector) => connector.id.as_str(),
+            Self::Plugin(plugin) => plugin.id.as_str(),
+        }
+    }
+
+    pub fn name(&self) -> &str {
+        match self {
+            Self::Connector(connector) => connector.name.as_str(),
+            Self::Plugin(plugin) => plugin.name.as_str(),
+        }
+    }
+
+    pub fn install_url(&self) -> Option<&str> {
+        match self {
+            Self::Connector(connector) => connector.install_url.as_deref(),
+            Self::Plugin(_) => None,
+        }
+    }
+}
+
+impl From<AppInfo> for DiscoverableTool {
+    fn from(value: AppInfo) -> Self {
+        Self::Connector(Box::new(value))
+    }
+}
+
+impl From<DiscoverablePluginInfo> for DiscoverableTool {
+    fn from(value: DiscoverablePluginInfo) -> Self {
+        Self::Plugin(Box::new(value))
+    }
+}
+
+pub fn filter_tool_suggest_discoverable_tools_for_client(
+    discoverable_tools: Vec<DiscoverableTool>,
+    app_server_client_name: Option<&str>,
+) -> Vec<DiscoverableTool> {
+    if app_server_client_name != Some(TUI_CLIENT_NAME) {
+        return discoverable_tools;
+    }
+
+    discoverable_tools
+        .into_iter()
+        .filter(|tool| !matches!(tool, DiscoverableTool::Plugin(_)))
+        .collect()
+}
+
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub struct DiscoverablePluginInfo {
+    pub id: String,
+    pub name: String,
+    pub description: Option<String>,
+    pub has_skills: bool,
+    pub mcp_server_names: Vec<String>,
+    pub app_connector_ids: Vec<String>,
 }
 
 #[derive(Clone, Debug, PartialEq, Eq)]
